requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
requests>=2.31.0
python-dateutil>=2.8.2

# API Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Dashboard & Visualization
streamlit>=1.25.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing (Optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Development (Optional)
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
